requirements.txt
# Core data & ML
pandas>=2.0
numpy>=1.24
scikit-learn>=1.3
scipy>=1.11
statsmodels>=0.14
pyyaml>=6.0

# Visualization
plotly>=5.10
kaleido==0.2.1

# MLflow
mlflow>=2.10

# API
fastapi>=0.110
uvicorn[standard]>=0.28
pydantic>=2.0

# Testing
pytest>=8.0
pytest-cov>=4.0

# Notebooks
jupyter>=1.0
nbconvert>=7.0
ipykernel>=6.0
